requirements.txt

streamlit
torch
transformers==4.38.2  # Use a stable version
faiss-cpu
numpy
pandas
sentence-transformers
sentencepiece
accelerate
evaluate
datasets
scipy
requests
beautifulsoup4
recipe-scrapers
PyPDF2
reportlab
